Refactor(senior-register): 시니어 등록 컴포넌트 분리
