+++ src/pages/guard/register/components/index.ts
@@ -1,1 +1,2 @@
 export * from './senior-info';
+export * from './senior-register-box';
