+++ src/pages/guard/register/components/senior-register-box/index.ts
@@ -1,0 +1,1 @@
+export { default as SeniorRegisterBox } from './SeniorRegisterBox';
